Add defaults tags for projects
Some projects are always used with the same tags. For instance, my
'python101' project should always have the 'teaching' tag attached to
it.

This commit adds a way to declare such default tags in the configuration
file, and have them applied. Default tags are expected to be declared in
a '[default_tags]' section:

    [default_tags]
    python101 = teaching
    watson = contrib python

With the sample configuration above, the 'teaching' tag will always be
attached to the 'python101' project, and the tags 'contrib' and 'python'
will always be attached to the project 'watson'.

Projects that do not appear in the '[default_tags]' section of the
configuration will not be affected by the change.

Note that the tags added by the new mechanism do not show in the output
of `watson start`. However, they appear in the output of `watson
status`, `watson log`, and of course `watson report`.