Add defaults tags for projects (#113)
This commit adds a way to declare default tags in the configuration file and have them applied on `watson start`. Default tags are declared in the `[default_tags]` section of the configuration:

    [default_tags]
    python101 = teaching
    watson = contrib python

With the sample configuration above, the "teaching" tag will always be attached to the "python101" project, and the tags "contrib" and "python" will always be attached to the project "watson". Tags in the configuration can also have spaces in them by enclosing them in single or double quotes.

    [default_tags]
    super project = fun 'random tag with spaces'

Or by using a multi-line option value (by indenting all but the first line) and listing each tag on its own line:

    [default_tags]
    python101 = teaching
        python book
        writing        
    watson = contrib python

Projects that do not appear in the `[default_tags]` section of the configuration will not be affected by the change.