Print log to terminal if the text fits
If the log fits in the terminal window, the log is printed to this
window. Elsewise, the log is given to the pager.

The terminal size is acquired using Linux command 'stty size', which
according to this post
(https://stackoverflow.com/questions/566746/how-to-get-linux-console-window-width-in-python)
is prevalent, so this should be widely compatible.