report/log: add lunar start time options
supplying --luna (e.g. instead of --month), the last full moon
will be taken as the start date of the summary.
The start date of a lunar month is not fixed, some cultures use
the new moon (chinese), some use the full moon (hindu) and some
use the first sight of the crescent (jewish / hebrew).

I implemented the full moon calendar because today is a full moon
and a total lunar eclipse.

My hope is to adjust my work load according the moon cycle.

Peace.