adds completion script for the fish shell
initial development work performed in https://github.com/andrewmeyer/fish-completions

expand documentation

expands documentation to include instructions on installing and
activating the fish completion script for watson.

provide lists of projects and tags for report/log

Extends the autocomplete functionality for log and report.
If -T/-p are provided, autocomplete a list of tags and projects