Don't stop frame when given `--seamless`
With this commit, `watson start --seamless` does not stop a started
frame, regardless of the configuration.

Using `--seamless` indicates that the user suspects that the stop time
of the youngest frame lies somewhat in the past. If this is not the
case, the user might want to use `watson stop --at` instead. If
`--seamless` would stop automatically, specifying the stop time would
then require two usages of `watson edit`.