Rename '--seamless' to '--no-gap'
The short option for the used keyword 'seamless', '-s' was used by
`watson restart` already. Furthermore, 'seamless' did not seem to be the
best choice.

On GitHub, '--nogap' was suggested, which seems to be far better. Thus,
everything regarding 'seamless' was renamed to 'no-gap' or 'gap',
depending on the context.