Remove `python2.7` from supported versions of Python
Version 2.7 of Python soon will reach the end of its live. It already
hinders the application of new language features which are incompatible
with 2.7. Thus, stopping to support `python2.7.` is necessary.

All major distributions ship some version of the Python3 series already.
Thus, stopping to support `python2.7` should be relatively light in
trouble for users.