Unify WatsonError catching in cli.py
This change has been motivated by the need to add tests for cli, because
watson.WatsonError was being redefined in cli.py which effectively broke
many existing tests in test_watson.py which relied on watson.WatsonError
not being WatsonCliError.

Besides this, some ValueError exceptions being raised in watson.py have
been changed to WatsonError in order to be properly formatted to the
user.