Allow to use fuzzed data in tests
- Change fuzzer.py to also support receving a path as argument where
fuzzed data will be generated. Without arguments, the behavior doesn't
change.
- New pytest fixture watson_df to support a non-empty config directory
using pytest-datafiles. This allows to work on a copy of the original
data files on each test.
- Create an empty directory where fuzzed data will be automatically
generated in CI, as specified in tox.ini.
- Make TEST_FIXTURE_DIR global in the tests module, as it will be
required by any test that wants to mark paths/files for
pytest-datafiles.

How to use the fuzzed data in a test through the new fixture:

    @pytest.mark.datafiles(TEST_FIXTURE_DIR / 'fuzzer_data')
    def test_something(watson_df):
        pass