BugFix Work around Click bug in all functions
The commit by David Alfonso improved the code style considerably.
Unfortunately it also triggered a bug in Click that does not pass the
context object properly.

While a workaround was added in some functions, some others have been
forgotten. This caused some attempts on autocompletion to fail.

The commit at hand applies the workaround everywhere where it is needed.