Update documentation of how to generate completion scripts
Now that completion scripts are distributed by the repository directly,
there is no need for the end-user documentation to cover how to generate
them.

However, in case newer click versions generate differing completion
scripts, when bumping a version the maintainer should be aware to
regenerate the completion scripts. Thus, a brief reminder to run
`make completion-scripts` is added there.